refactor: 패키지 구조 변경 및 불필요한 메서드 삭제
