+++ src/main/java/blackjack/controller/DefaultGame.java
@@ -1,8 +1,8 @@
 package blackjack.controller;
 
 import blackjack.domain.participant.Dealer;
-import blackjack.domain.participant.Name;
 import blackjack.domain.participant.Players;
+import blackjack.domain.participant.attribute.Name;
 import blackjack.domain.result.PlayersResults;
 import blackjack.domain.result.ResponseDTO.WinningDTO;
 import blackjack.view.InputView;
