+++ src/main/java/blackjack/domain/card/CardFactory.java
@@ -1,5 +1,8 @@
 package blackjack.domain.card;
 
+import blackjack.domain.card.component.Figure;
+import blackjack.domain.card.component.Type;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
