+++ src/main/java/blackjack/domain/card/component/Type.java
@@ -1,4 +1,4 @@
-package blackjack.domain.card;
+package blackjack.domain.card.component;
 
 public enum Type {
     TWO(2, "2"),
