+++ src/test/java/blackjack/domain/card/CardsTest.java
@@ -1,5 +1,7 @@
 package blackjack.domain.card;
 
+import blackjack.domain.card.component.Figure;
+import blackjack.domain.card.component.Type;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
