+++ src/test/java/blackjack/domain/result/PlayersResultsTest.java
@@ -1,9 +1,9 @@
 package blackjack.domain.result;
 
 import blackjack.domain.participant.BettingPlayer;
-import blackjack.domain.participant.Money;
-import blackjack.domain.participant.Name;
 import blackjack.domain.participant.Player;
+import blackjack.domain.participant.attribute.Money;
+import blackjack.domain.participant.attribute.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
